Insert vehicles, axle chart and axles with new permit

+++ src/main/java/fi/vaylavirasto/sillari/model/AxleModel.java
@@ -2,12 +2,14 @@ package fi.vaylavirasto.sillari.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AxleModel {
     private Integer id;
     private Integer axleChartId;
     private Integer axleNumber;
-    private Double weight;
-    private Double distanceToNext;
-    private Double maxDistanceToNext;
+    private BigDecimal weight;
+    private BigDecimal distanceToNext;
+    private BigDecimal maxDistanceToNext;
 }
